create separated custom selects for persons, relation types and location instances
